feat(core): implement modular hash parsing system
 • Refactor StorageHash interface and related functions into new core/storage_hash.go
 • Introduce StorageHashParser interface for extensible hash format parsing
 • Add core parsers for Base58, Hex, Base64, and Base32 multihash formats
 • Implement parser registry that combines protocol-specific and core parsers
 • Improve CID type inference logic for different multihash configurations

This change improves code organization and enables easier addition of new hash formats while maintaining backward compatibility with existing multihash parsing functionality.
